requirements.txt
keras
numpy
pillow
tensorflow
imageio
